+++ src/test/java/it/cnr/ilc/lexo/manager/LexicalEntryManagerTest.java
@@ -119,6 +119,8 @@ class LexicalEntryManagerTest {
 
             assertThat(connection.hasStatement(entry, iri(ONTOLEX + "sense"),
                     sense, false, graph)).isTrue();
+            assertThat(connection.hasStatement(sense,
+                    iri(ONTOLEX + "isSenseOf"), entry, false, graph)).isTrue();
             assertThat(connection.hasStatement(sense, RDF.TYPE,
                     iri(ONTOLEX + "LexicalSense"), false, graph)).isTrue();
             assertThat(connection.hasStatement(sense, iri(SKOS + "definition"),